backend/requirements.txt
# ==============================
# Core Django
# ==============================
Django>=4.2,<5.0

# ==============================
# WSGI Server (Production)
# ==============================
gunicorn>=21.2

# ==============================
# Database (PostgreSQL)
# ==============================
psycopg2-binary>=2.9

# ==============================
# REST Framework
# ==============================
djangorestframework>=3.14

# ==============================
# CORS Handling
# ==============================
django-cors-headers>=4.3

# ==============================
# Static File Serving (Docker)
# ==============================
whitenoise>=6.6

# ==============================
# Environment Variable Handling
# ==============================
python-dotenv>=1.0

# ==============================
# Dev / Utility (선택)
# ==============================
django-extensions>=3.2
